Edit participants

See merge request cse1105/2023-2024/teams/oopp-team-56!94

+++ server/src/test/java/server/api/TestQuoteRepository.java
@@ -141,7 +141,7 @@ public class TestQuoteRepository implements QuoteRepository {
     @Override
     public <S extends Quote> S save(S entity) {
         call("save");
-        entity.id = (long) quotes.size();
+        entity.id = quotes.size();
         quotes.add(entity);
         return entity;
     }
